Combine budget-group w/dash-budgets to make 1 comp
budget-group comp ended up being too difficult to manage as its own
component, so they were merged into one
