--- src/app/components/budget-group/budget-group.component.css
@@ -1,25 +1,0 @@
-h3 {
-  /* self style */
-  font-weight: 500;
-}
-
-table {
-  /* self align */
-  margin-bottom: 20px;
-}
-
-.mat-column-amountLimit {
-  width: 10%;
-}
-
-.mat-column-period {
-  width: 20%;
-}
-
-.mat-column-startDate {
-  width: 20%;
-}
-
-.mat-column-currentPeriod {
-  width: 30%;
-}
