--- src/app/components/budget-group/budget-group.component.cy.ts
@@ -1,7 +1,0 @@
-import { BudgetGroupComponent } from './budget-group.component'
-
-describe('BudgetGroupComponent', () => {
-  it('should mount', () => {
-    cy.mount(BudgetGroupComponent)
-  })
-})
