+++ src/app/components/dash-budgets/dash-budgets.component.ts
@@ -21,6 +21,17 @@ export class DashBudgetsComponent {
   budgetForm!: FormGroup;
   budgetId?: number;
   mode!: "Add" | "Edit";
+  numberFormatter = new Intl.NumberFormat('en-US', {
+    style: 'currency',
+    currency: 'USD'
+  })
+  displayedColumns = [
+    "amountLimit",
+    "period",
+    "startDate",
+    "currentPeriod",
+    "editAndDelete"
+  ]
 
   constructor(
     private budgetService: BudgetService,
@@ -41,8 +52,107 @@ export class DashBudgetsComponent {
               this.existingCategories.push(elem.data.category);
             }
           })
+          this.existingCategories.sort();
         }
+        console.log(this.budgetData);
       })
+  }
+
+  getPeriodDef(budgetContent: BudgetContent) {
+    const frequency = budgetContent.frequency;
+    const duration = budgetContent.duration;
+    const count = budgetContent.count;
+    let result = "Every ";
+    let frequencyString =
+      frequency === Period.monthly
+        ? "month"
+        : frequency === Period.weekly
+          ? "week"
+          : frequency === Period.yearly
+            ? "year"
+            : "period"
+    if (duration > 1) {
+      result += `${duration} ${frequencyString}s`;
+    } else {
+      result += `${frequencyString}`;
+    }
+
+    if (count) {
+      result += ` for ${count} ${frequencyString}`;
+      if (count > 1) {
+        result += 's';
+      }
+    }
+    return result;
+  }
+
+  parseDate(str: string) {
+    return new Date(str).toLocaleDateString();
+  }
+
+  getPeriodString(budgetContent: BudgetContent) {
+    const currentPeriod = this.getPeriod(budgetContent);
+    if (currentPeriod === null) {
+      return "Expired";
+    }
+    return `${currentPeriod.periodStart.toLocaleDateString()} 
+    to ${currentPeriod.periodEnd.toLocaleDateString()}
+    (${currentPeriod.daysLeft} day(s) left)`
+  }
+
+  getPeriod(budgetContent: BudgetContent): { periodStart: Date, periodEnd: Date, daysLeft: number } | null {
+    const startDate = new Date(budgetContent.startDate);
+    const today = new Date();
+
+    let addOne: (current: Date) => Date;
+    if (budgetContent.frequency === Period.monthly) {
+      addOne = (current) => {
+        current.setMonth(current.getMonth() + 1);
+        return current;
+      }
+    } else if (budgetContent.frequency === Period.yearly) {
+      addOne = (current) => {
+        current.setFullYear(current.getFullYear() + 1);
+        return current;
+      }
+    } else if (budgetContent.frequency === Period.weekly) {
+      addOne = (current) => {
+        current.setDate(current.getDate() + 7);
+        return current;
+      }
+    } else {
+      addOne = (current) => {
+        current.setDate(current.getDate() + 1);
+        return current;
+      }
+    }
+
+    let periodStart = new Date(startDate);
+    let periodEnd = new Date(startDate);
+    for (let i = 0; i < budgetContent.duration; i++) {
+      periodEnd = addOne(periodEnd);
+    }
+
+
+    let periodsPassed = 0;
+
+    while (periodEnd < today && (!budgetContent.count || periodsPassed < budgetContent.count)) {
+      for (let i = 0; i < budgetContent.duration; i++) {
+        periodStart = addOne(periodStart);
+        periodEnd = addOne(periodEnd);
+      }
+      periodsPassed++;
+
+    }
+
+    if (budgetContent.count && periodsPassed >= budgetContent.count) {
+      return null;
+    }
+    return ({
+      periodStart: periodStart,
+      periodEnd: periodEnd,
+      daysLeft: Math.floor((periodEnd.getTime() - today.getTime()) / 86400000)
+    })
   }
 
   getFilteredData(category: string, budgetData: Budget[]) {
@@ -66,10 +176,12 @@ export class DashBudgetsComponent {
       } as BudgetsDialogData
     });
 
-    dialogRef.afterClosed().subscribe((res?: BudgetsDialogData) => {
+    dialogRef.afterClosed().subscribe((res?: Budget) => {
+
       if (res) {
+
         const budgetRequest: CreateBudgetRequest = {
-          ...res.data.data
+          ...res.data
         }
         this.budgetService.createBudget(budgetRequest).subscribe(_ => {
           this.rerenderBudgets();
@@ -87,10 +199,10 @@ export class DashBudgetsComponent {
       } as BudgetsDialogData
     });
 
-    dialogRef.afterClosed().subscribe((res?: BudgetsDialogData) => {
+    dialogRef.afterClosed().subscribe((res?: Budget) => {
       if (res) {
         const budgetRequest: UpdateBudgetRequest = {
-          newBudget: res.data,
+          newBudget: res
         }
         this.budgetService.updateBudget(budgetRequest).subscribe(_ => {
           this.rerenderBudgets();
@@ -144,15 +256,19 @@ export class BudgetsDialogComponent {
     if (!this.budgetForm.invalid) {
 
       const repeats: boolean = this.budgetForm.get('repeats')?.value as boolean
-      const budgetContent: BudgetContent = {
-        category: this.budgetForm.get('category')?.value as string,
-        amountLimit: this.budgetForm.get('amount')?.value as number,
-        frequency: this.budgetForm.get('frequency')?.value as Period,
-        duration: this.budgetForm.get('duration')?.value as number,
-        count: repeats ? this.budgetForm.get('count')?.value as number : undefined,
-        startDate: this.budgetForm.get('startDate')?.value as string,
+      const budget: Budget = {
+        userId: this.data.data.userId,
+        budgetId: this.data.data.budgetId,
+        data: {
+          category: this.budgetForm.get('category')?.value as string,
+          amountLimit: this.budgetForm.get('amount')?.value as number,
+          frequency: this.budgetForm.get('frequency')?.value as Period,
+          duration: this.budgetForm.get('duration')?.value as number,
+          count: repeats ? this.budgetForm.get('count')?.value as number : undefined,
+          startDate: this.budgetForm.get('startDate')?.value as string,
+        },
       }
-      this.dialogRef.close(budgetContent);
+      this.dialogRef.close(budget);
     }
   }
